Tested and fixed services
- All services behave correctly (apart of the command service)
- Minor code cleanup and comments
- Added a maintenance controller that can manually progress turns
- Added some global values in KnownValues

+++ backend/StrategyGame.Bll/DTO/Sent/Country/CountryInfo.cs
@@ -10,7 +10,7 @@ namespace StrategyGame.Bll.Dto.Sent.Country
     /// </summary>
     public class CountryInfo
     {
-        public int Round { get; set; }
+        public ulong Round { get; set; }
         public int Rank { get; set; }
         public IEnumerable<UnitInfo> ArmyInfo { get; set; }
         public long Pearls { get; set; }
